Add circular style dashboard view
