+++ Example/Podfile
@@ -1,6 +1,7 @@
 use_frameworks!
 
 target 'Example' do
+
   pod 'LifetimeTracker', :path => '../'
 
   target 'Example_Tests' do
