+++ Sources/Extensions.swift
@@ -10,10 +10,10 @@ import Foundation
 
 internal extension String {
 
-	var lt_localized: String {
+    var lt_localized: String {
         guard let path = Bundle(for: LifetimeTracker.self).path(forResource: "LifetimeTracker", ofType: "bundle"), let bundle = Bundle(path: path) else {
             return self
         }
-		return NSLocalizedString(self, bundle: bundle, comment: self)
-	}
+        return NSLocalizedString(self, bundle: bundle, comment: self)
+    }
 }
